+++ mw/base.ts
@@ -16,7 +16,8 @@ void registerWiki(monaco, true);
 mw.loader.load(`${CDN}/monaco-editor/esm/vs/base/browser/ui/codicons/codicon/codicon.min.css`, 'text/css');
 mw.loader.addStyleTag(
 	'.monaco-editor .glyph-margin-widgets>.codicon-warning::before{color:var(--vscode-problemsWarningIcon-foreground)}'
-	+ '.monaco-editor .glyph-margin-widgets>.codicon-error::before{color:var(--vscode-problemsErrorIcon-foreground)}',
+	+ '.monaco-editor .glyph-margin-widgets>.codicon-error::before{color:var(--vscode-problemsErrorIcon-foreground)}'
+	+ '.monaco-hover-content code{color:inherit}',
 );
 
 class MonacoWikiEditor {